# Faster union find via unsafe array indexing