Before:
```
test regionlabelling::test::bench_connected_components_eight_chessboard ... bench:   2,449,970 ns/iter (+/- 651,403)
test regionlabelling::test::bench_connected_components_four_chessboard  ... bench:   7,818,751 ns/iter (+/- 1,835,232)
```
After:

```
test regionlabelling::test::bench_connected_components_eight_chessboard ... bench:   2,168,510 ns/iter (+/- 227,603)
test regionlabelling::test::bench_connected_components_four_chessboard  ... bench:   6,421,001 ns/iter (+/- 733,899)
```

This is massively understating the speed up to union find itself, as currently a huge chunk of the time in connected_components is spent relabelling the regions to start at 1 (TODO: fix that!).